% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp3.R
\name{exp3}
\alias{exp3}
\title{Exponential-weight algorithm for Exploration and Exploitation Policy}
\usage{
exp3(k, gamma = 0.05)
}
\arguments{
\item{k}{number of arms}

\item{gamma}{exploration parameter}
}
\value{
An agent object, i.e. a list of two functions : choose and receive
}
\description{
Exponential-weight algorithm for Exploration and Exploitation Policy
}
