% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp3.R
\name{expp}
\alias{expp}
\title{EXP3 expectation estimation of an arm}
\usage{
expp(w, sum, k, gamma)
}
\arguments{
\item{w}{arm's weight}

\item{sum}{sum of all arm weights}

\item{k}{number of arms}

\item{gamma}{exploration parameter}
}
\value{

}
\description{
EXP3 expectation estimation of an arm
}
