% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_distribution.R
\name{from_distribution}
\alias{from_distribution}
\title{Fixed arm pull probabilities Policy}
\usage{
from_distribution(k, prob)
}
\arguments{
\item{k}{number of arms}

\item{prob}{pull probability of each arm}
}
\value{
An agent object, i.e. a list of two functions : choose and receive
}
\description{
Fixed arm pull probabilities Policy
}
