% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thompson_sampling.R
\name{thompson_sampling}
\alias{thompson_sampling}
\title{Thompson Sampling Policy}
\usage{
thompson_sampling(k, alpha = 1, beta = 1)
}
\arguments{
\item{k}{number of arms}

\item{alpha}{first beta distribution parameter}

\item{beta}{second beta distribution parameter}
}
\value{
An agent object, i.e. a list of two functions : choose and receive
}
\description{
Thompson Sampling Policy
}
