% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thompson_sampling.R
\name{ts}
\alias{ts}
\title{Thompson sample value of an arm}
\usage{
ts(mu, nu, alpha, beta)
}
\arguments{
\item{mu}{observed mean}

\item{nu}{number of trials}

\item{alpha}{first beta distribution parameter}

\item{beta}{second beta distribution parameter}
}
\value{
double value
}
\description{
Thompson sample value of an arm
}
