% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_confidence_bound.R
\name{ucb}
\alias{ucb}
\title{Upper Confidence Bound value of an arm}
\usage{
ucb(mu, nu, alpha, t)
}
\arguments{
\item{mu}{observed mean}

\item{nu}{number of trials}

\item{alpha}{exploration parameter}

\item{t}{iteration index}
}
\value{
double value
}
\description{
Upper Confidence Bound value of an arm
}
