% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_confidence_bound.R
\name{upper_confidence_bound}
\alias{upper_confidence_bound}
\title{Upper Confidence Bound Policy}
\usage{
upper_confidence_bound(k, alpha = 1)
}
\arguments{
\item{k}{number of arms}

\item{alpha}{exploration parameter}
}
\value{
An agent object, i.e. a list of two functions : choose and receive
}
\description{
Upper Confidence Bound Policy
}
